+++ catalog/app/utils/BucketPreferences/Provider.tsx
@@ -1,36 +1,51 @@
-import * as R from 'ramda'
+import type { S3 } from 'aws-sdk'
 import * as React from 'react'
 
+import type * as Model from 'model'
 import cfg from 'constants/config'
 import * as quiltConfigs from 'constants/quiltConfigs'
-import * as bucketErrors from 'containers/Bucket/errors'
+import { FileNotFound, VersionNotFound } from 'containers/Bucket/errors'
 import * as requests from 'containers/Bucket/requests'
 import * as AWS from 'utils/AWS'
 import * as CatalogSettings from 'utils/CatalogSettings'
 import { useData } from 'utils/Data'
 
-import { Result, parse } from './BucketPreferences'
+import {
+  BucketPreferences,
+  BucketPreferencesInput,
+  Result,
+  merge,
+  openInDesktop,
+  parse,
+  validate,
+} from './BucketPreferences'
 import LocalProvider from './LocalProvider'
 
 interface FetchBucketPreferencesArgs {
-  s3: $TSFixMe
+  s3: S3
   bucket: string
+  counter: number
 }
 
-async function fetchBucketPreferences({ s3, bucket }: FetchBucketPreferencesArgs) {
+interface FetchBucketPreferencesOutput {
+  body: string
+  handle: Model.S3.S3ObjectLocation | null
+}
+
+async function fetchBucketPreferences({
+  s3,
+  bucket,
+}: FetchBucketPreferencesArgs): Promise<FetchBucketPreferencesOutput> {
   try {
-    const response = await requests.fetchFile({
+    const { handle, body } = await requests.fetchFileInCollection({
       s3,
-      bucket,
-      path: quiltConfigs.bucketPreferences,
+      handles: quiltConfigs.bucketPreferences.map((key) => ({ bucket, key })),
     })
-    return parse(response.Body.toString('utf-8'))
+    return { body: body?.toString('utf-8') || '', handle }
   } catch (e) {
-    if (
-      e instanceof bucketErrors.FileNotFound ||
-      e instanceof bucketErrors.VersionNotFound
-    )
-      return parse('')
+    if (e instanceof FileNotFound || e instanceof VersionNotFound) {
+      return { body: '', handle: null }
+    }
 
     // eslint-disable-next-line no-console
     console.log('Unable to fetch')
@@ -40,24 +55,83 @@ async function fetchBucketPreferences({ s3, bucket }: FetchBucketPreferencesArgs
   }
 }
 
-const Ctx = React.createContext<Result>(Result.Init())
+async function uploadBucketPreferences(
+  s3: S3,
+  bucket: string,
+  update: BucketPreferencesInput,
+) {
+  const response = await fetchBucketPreferences({
+    s3,
+    bucket,
+    counter: 1,
+  })
+  const updatedConfig = merge(response.body, update)
+  const handle = response.handle || { bucket, key: quiltConfigs.bucketPreferences[0] }
+
+  validate(updatedConfig)
+
+  await s3
+    .putObject({
+      Bucket: handle.bucket,
+      Key: handle.key,
+      Body: updatedConfig,
+    })
+    .promise()
+  return parse(updatedConfig)
+}
+
+interface State {
+  handle: Model.S3.S3ObjectLocation | null
+  prefs: Result
+  update: (upd: BucketPreferencesInput) => Promise<BucketPreferences>
+}
+
+const Ctx = React.createContext<State>({
+  handle: null,
+  prefs: Result.Init(),
+  update: () => Promise.reject(new Error('Bucket preferences context not initialized')),
+})
 
 type ProviderProps = React.PropsWithChildren<{ bucket: string }>
 
 function CatalogProvider({ bucket, children }: ProviderProps) {
   const s3 = AWS.S3.use()
   const settings = CatalogSettings.use()
-  const data = useData(fetchBucketPreferences, { s3, bucket })
+  const [counter, setCounter] = React.useState(0)
+  const data = useData(fetchBucketPreferences, { s3, bucket, counter })
 
-  const preferences = data.case({
-    Ok: settings?.beta
-      ? R.pipe(R.assocPath(['ui', 'actions', 'openInDesktop'], true), Result.Ok)
-      : Result.Ok,
+  const update = React.useCallback(
+    async (upd: BucketPreferencesInput) => {
+      const preferences = await uploadBucketPreferences(s3, bucket, upd)
+      setCounter((prev) => prev + 1)
+      return preferences
+    },
+    [s3, bucket],
+  )
+
+  const prefs = data.case({
+    Ok: ({ body }: FetchBucketPreferencesOutput) => {
+      try {
+        const input = settings?.beta ? merge(body, openInDesktop()) : body
+        return Result.Ok(parse(input))
+      } catch (e) {
+        // eslint-disable-next-line no-console
+        console.log('Unable to parse bucket preferences')
+        // eslint-disable-next-line no-console
+        console.error(e)
+        return Result.Ok(parse(''))
+      }
+    },
     Err: () => Result.Ok(parse('')),
     Pending: Result.Pending,
     Init: Result.Init,
   })
-  return <Ctx.Provider value={preferences}> {children} </Ctx.Provider>
+  const handle = data.case({
+    Ok: (r: FetchBucketPreferencesOutput) => r.handle,
+    _: () => null,
+  })
+
+  return <Ctx.Provider value={{ handle, prefs, update }}> {children} </Ctx.Provider>
 }
 
 export function Provider({ bucket, children }: ProviderProps) {